Use long instead of int for event IDs
Also make User.SaltedPasswordHash public

+++ EvernestFront/EvernestFrontTests/SourceTests.cs
@@ -51,7 +51,7 @@ namespace EvernestFrontTests
             Assert.IsNull(ans.Error);
         }
 
-        internal static int GetEventId_AssertSuccess(string sourceKey, String message)
+        internal static long GetEventId_AssertSuccess(string sourceKey, String message)
         {
             Source source = GetSource_AssertSuccess(sourceKey);
             Push ans = source.Push(message);
@@ -186,8 +186,8 @@ namespace EvernestFrontTests
             var userId = UserTests.AddUser_GetId_AssertSuccess(UserName);
             long stream = UserTestsTowardEventStream.CreateEventStream_GetId_AssertSuccess(userId, StreamName);
             var sourceKey = CreateSource_GetKey_AssertSuccess(userId, stream, SourceName, AccessRights.Admin);
-            int eventId = GetEventId_AssertSuccess(sourceKey, Message);
-            int eventId2 = GetEventId_AssertSuccess(sourceKey, Message);
+            long eventId = GetEventId_AssertSuccess(sourceKey, Message);
+            long eventId2 = GetEventId_AssertSuccess(sourceKey, Message);
             Assert.AreEqual(eventId+1, eventId2);
         }
 
@@ -224,7 +224,7 @@ namespace EvernestFrontTests
             long stream = UserTestsTowardEventStream.CreateEventStream_GetId_AssertSuccess(userId, StreamName);
             var sourceKey = CreateSource_GetKey_AssertSuccess(userId, stream, SourceName, AccessRights.ReadOnly);
             Source source = GetSource_AssertSuccess(sourceKey);
-            int eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(userId, stream, Message);
+            long eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(userId, stream, Message);
             var ans = source.Pull(eventId);
             Assert.IsTrue(ans.Success);
             Assert.IsNull(ans.Error);
@@ -242,7 +242,7 @@ namespace EvernestFrontTests
             var userId = UserTests.AddUser_GetId_AssertSuccess(UserName);
             var sourceKey = CreateSource_GetKey_AssertSuccess(userId, stream, SourceName, AccessRights.ReadOnly); //user has no rights on stream
             Source source = GetSource_AssertSuccess(sourceKey);
-            int eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(creatorId, stream, Message);
+            long eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(creatorId, stream, Message);
             var ans = source.Pull(eventId);
             AssertAuxiliaries.ErrorAssert<ReadAccessDenied>(ans);
         }
@@ -254,7 +254,7 @@ namespace EvernestFrontTests
             long stream = UserTestsTowardEventStream.CreateEventStream_GetId_AssertSuccess(userId, StreamName);
             var sourceKey = CreateSource_GetKey_AssertSuccess(userId, stream, SourceName, AccessRights.WriteOnly); //source cannot read
             Source source = GetSource_AssertSuccess(sourceKey);
-            int eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(userId, stream, Message);
+            long eventId = UserTestsTowardEventStream.PushEvent_GetId_AssertSuccess(userId, stream, Message);
             var ans = source.Pull(eventId);
             AssertAuxiliaries.ErrorAssert<ReadAccessDenied>(ans);
         }
